Move search to compnents folder
